requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.22.0
tiktoken==0.4.0
protobuf~=3.19.0
langchain_experimental
sentence-transformers
chromadb
google-generativeai
langchain-google-genai
SQLAlchemy